+++ Holovibes/sources/api/globalpostprocess_api.cc
@@ -3,6 +3,16 @@
 namespace holovibes::api
 {
 
+#pragma region Renrm
+
+void set_renorm_enabled(bool value)
+{
+    UPDATE_SETTING(RenormEnabled, value);
+    pipe_refresh();
+}
+
+#pragma endregion
+
 #pragma region Registration
 
 void update_registration_zone(float value)
@@ -12,6 +22,18 @@ void update_registration_zone(float value)
 
     set_registration_zone(value);
     api::get_compute_pipe()->request(ICS::UpdateRegistrationZone);
+    pipe_refresh();
+}
+
+void set_registration_enabled(bool value)
+{
+    if (api::get_compute_mode() == Computation::Raw)
+        return;
+
+    if (!api::get_fft_shift_enabled())
+        set_fft_shift_enabled(value);
+
+    UPDATE_SETTING(RegistrationEnabled, value);
     pipe_refresh();
 }
